+++ core/spec/spec_helper.rb
@@ -1,0 +1,5 @@
+require 'torquebox'
+
+def jruby_command
+  File.join(RbConfig::CONFIG['bindir'], RbConfig::CONFIG['ruby_install_name'])
+end