+++ web/lib/torquebox-web.rb
@@ -1,0 +1,9 @@
+Dir.glob("#{File.dirname(__FILE__)}/wunderboss-jars/*.jar") do |jar|
+  require jar
+end
+
+require 'wunderboss-rack.jar'
+require 'wunderboss_rack_response_handler'
+
+require 'torquebox/web/cli'
+require 'torquebox/server'